Base Linux functions
This is what I managed to come up with after a bit of fiddling. I still need to figure out how to get the mouse clicks, but I think I have a solution in mind. This proposed file definitely needs to be cleaned up and some things modified before merging.