Create Click.py
This class will have to be called when it comes time to record all the click locations. It places them in a log file that also will be created. It should be called like this: 

`from Click import mouse_click
get_mouse_click = mouse_click('insert/path/here')`

If anybody has a better way to do it I'd love to know.