Program detection in main thread didn't work
The idea was for the background thread to parse the program list, and the foreground thread to check it, so that during processing heavy situations it'll still detect programs.
It worked, but unfortunately I hadn't realised the background thread would still be required to parse the task list, which it couldn't do if it was doing heavy processing. An alternative would be to start a third process, which I may look into later.