Fixed rare crash on unusual resolutions
In the multi-monitor part of code, it would only recalculate offsets if there was a change in resolution. However there has been a typo that's existed for months that caused these values not to be set, and it was a massive effort to figure out. Also slightly optimised when the offset function is called.