Changed .data extension to .mtk
The MTK extension doesn't appear to be in common use, so it can be safely set to generate images if the file is double clicked. I decided against putting in code to check if the older .data file exists, as the script is not released yet and the files are easy to rename.