Fix console show/hide issues
It was hard to figure this one out. It seems to require a new integrated console in Visual Studio Code for the bug to happen.
The integrated consoles use conhost.exe, which requires its own hidden process to run.

On the first script launch in the new integrated console, the PID can be read (not sure why, but not important).
This then enabled the show/hide features, and this hidden window would then load if an error occurred.

From my observations, an invalid console is the only one that's hidden on startup, and it doesn't have any title.
Resolves #74.