Create the first release
This is the first official public release of MouseTracks!

After originally starting work on the project in 2017, then picking it up again in early 2025, I'm excited to say it's finally ready to see the light of day.
All the core features are in place, with a lot of effort put into making sure the application runs as smoothly and as reliably as possible.

# Features
- Live activity tracking
- Per-application profiling
- High quality rendering
- Customisable artistic renders
- Multi-monitor support
- Interactive live GUI
- Import/export data
- Standalone application (no install required)
- Option to launch on startup